common base class for sets, how to handle windowing
